Phase 5: Move and rename ingress (fix router misnomer)
- Move deployments/routers → core/ingress
- Rename router.py → ingress.py (KEY FIX)
- Rename builder_ingress.py → builder.py
- Update all imports
