+++ python/ray/llm/tests/serve/cpu/deployments/routers/test_builder_ingress.py
@@ -14,12 +14,12 @@ from ray.llm._internal.serve.core.configs.llm_config import (
     LLMConfig,
     ModelLoadingConfig,
 )
-from ray.llm._internal.serve.deployments.routers.builder_ingress import (
+from ray.llm._internal.serve.core.ingress.builder import (
     IngressClsConfig,
     LLMServingArgs,
     build_openai_app,
 )
-from ray.llm._internal.serve.deployments.routers.router import OpenAiIngress
+from ray.llm._internal.serve.core.ingress.ingress import OpenAiIngress
 from ray.serve.config import AutoscalingConfig
